requirements.txt
httpx>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.7.0
InquirerPy>=0.3.4
openai>=1.0.0
dnspython>=2.4.0
